# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PDF Alchemist",
        "description": "Swagger dokumentácia pre všetky REST endpointy PDF aplikácie.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://zadanie-skuska.test",
            "description": "Lokálny vývojový server"
        }
    ],
    "paths": {
...
        "/api/api-key": {
            "get": {
                "tags": [
                    "API Key"
                ],
                "summary": "Generate a new API key",
                "description": "Generates a new API key for the currently authenticated user and returns it in plain text. The key is stored in hashed form in the database.",
                "operationId": "0db5088d83da4e8f174520aac3df03a4",
                "responses": {
                    "200": {
                        "description": "API key generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "api_key": {
                                            "type": "string",
                                            "example": "9e2cf4f879d449a89bfb1b44b196cf6b0f0a24e86d7356c7df4832e2d49956a7"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized – user must be authenticated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Session-based login",
                "operationId": "31dd365bddb46a996c2350b0a1fb7381",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user (session-based)",
                "operationId": "392091b67067bb84702403a281d29dff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "username",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration successful"
                    }
                }
            }
        },
...
        "/api/download/{token}": {
            "get": {
                "tags": [
                    "Download"
                ],
                "summary": "Download file using temporary token",
                "description": "Downloads a file if a valid session token exists.",
                "operationId": "e8010ac81df74068fcbb009ccdbda80b",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Temporary download token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123xyz"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File download (binary)"
                    },
                    "404": {
                        "description": "File not found or token expired"
                    }
                }
            }
        },
        "/api/manual": {
            "get": {
                "tags": [
                    "Manual"
                ],
                "summary": "Get user manual HTML content",
                "description": "Returns rendered HTML content of the localized user manual.",
                "operationId": "28c1ae784ea8a130b4eca5b6e34fab98",
                "responses": {
                    "200": {
                        "description": "Manual HTML loaded successfully"
                    },
                    "404": {
                        "description": "Manual not found"
                    },
                    "500": {
                        "description": "Rendering error"
                    }
                }
            }
        },
        "/api/manual/pdf": {
            "get": {
                "tags": [
                    "Manual"
                ],
                "summary": "Download user manual as PDF",
                "description": "Returns the user manual as a PDF file for the current locale.",
                "operationId": "6d3e17ebcaea703e803c87ac10771c4a",
                "responses": {
                    "200": {
                        "description": "PDF manual downloaded"
                    },
                    "404": {
                        "description": "Manual not found"
                    },
                    "500": {
                        "description": "PDF generation error"
                    }
                }
            }
        },
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get current authenticated user profile",
                "operationId": "650b2420492dafd00dec1efd87fb19e8",
                "responses": {
                    "200": {
                        "description": "Authenticated user profile returned"
                    }
                }
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile of authenticated user",
                "operationId": "cb8d53629911073b19c0914694bbed73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
...
    },
...
}